Use PackedRead::read_octetstring for OctetString and Utf8String

+++ src/syn/octetstring.rs
@@ -10,8 +10,8 @@ impl<C: Constraint> Default for OctetString<C> {
 }
 
 pub trait Constraint {
-    const MIN: Option<usize> = None;
-    const MAX: Option<usize> = None;
+    const MIN: Option<u64> = None;
+    const MAX: Option<u64> = None;
     const EXTENSIBLE: bool = false;
 }
